fading_restoration/requirements.txt
Pillow==9.3.0
torch==1.13.1
torchvision==0.14.1
numpy==1.23.5
Flask==2.0.2
Werkzeug==2.0.1
